Refactored logging and testing infrastructure, added new ApiTesting namespace.
This commit includes significant changes to the testing and logging infrastructure. The improvements resulted in cleaner API for test logging, allowing multiple kinds of test loggers such as PulseFlow and SimpleTestLogger. Moreover, a new namespace, ApiTesting, was established which includes Assertion, AssertionGroup and Result classes. This will enhance the organization and readability of test cases. This commit also included relevant changes to the project and solution files.