Update package references and add conditional packing
The commit revises the Directory.Build.props to only include README.md and icon.png if they don't already exist. It also updates the version of Frank.Reflection in the Frank.Testing.Logging project and replaces VarDump with Frank.Reflection.Dump in Frank.Testing.TestOutputExtensions. The commit introduces a new function in TestOutputCSharpExtensions to handle output of an Enumerable source.