Update test base classes and upgrade package versions
The key changes in this commit are focused on improving the test base classes and upgrading software package versions. The update to the loggers introduced in the WebApplicationTestBase and HostApplicationTestBase classes enhances the allowed variability for loggers and providers. The upgrades include the newest versions of Microsoft.NET.Test.Sdk, xunit, coverlet.collector and Testcontainers, essentially ensuring compatibility with the latest libraries. Some outdated and unused test components were also removed for better maintainability.